+++ packages/activity_recognition_flutter/lib/src/ar_channel.dart
@@ -6,13 +6,21 @@
 part of activity_recognition;
 
 class _ActivityChannel {
-  StreamController<Activity> _activityStreamController =
-      StreamController<Activity>();
+  StreamController<ActivityEvent> _activityStreamController =
+      StreamController<ActivityEvent>();
   StreamSubscription _activityUpdateStreamSubscription;
 
-  Stream<Activity> get activityUpdates => _activityStreamController.stream;
+  Stream<ActivityEvent> get activityUpdates => _activityStreamController.stream;
 
-  _ActivityChannel() {
+  bool _runForegroundService;
+
+  _ActivityChannel(this._runForegroundService) {
+    /// Start the foreground service plugin if we are on android
+    /// and if the option was not toggled off by the programmer.
+    if (Platform.isAndroid && _runForegroundService) {
+      ForegroundService().start();
+    }
+
     _activityStreamController.onListen = startActivityUpdates();
   }
 
@@ -36,7 +44,7 @@ class _ActivityChannel {
   _onActivityUpdateReceived(dynamic activity) {
     debugPrint("onActivityUpdateReceived");
     assert(activity is String);
-    var parsedActivity = Activity.fromJson(json.decode(activity));
+    var parsedActivity = ActivityEvent.fromJson(json.decode(activity));
     _activityStreamController.add(parsedActivity);
   }